# Parse IPv6 without net.ParseIP, and make it faster.