With this change, netaddr is no longer dependent on net.ParseIP for its
    ParseIP implementation. It's also 2-4x faster than net.ParseIP and
    allocation-free, except for v6+zone addresses which are still slower and
    require 1 allocation.

```    
    name                      old time/op    new time/op    delta
    ParseIP/v4-8                22.0ns ± 2%    24.5ns ± 1%   +11.46%  (p=0.016 n=5+4)
    ParseIP/v6-8                 169ns ± 2%      91ns ± 2%   -45.97%  (p=0.008 n=5+5)
    ParseIP/v6_ellipsis-8        140ns ± 2%      68ns ± 5%   -51.32%  (p=0.008 n=5+5)
    ParseIP/v6_v4-8              216ns ± 1%      71ns ± 1%   -67.19%  (p=0.008 n=5+5)
    ParseIP/v6_zone-8            526ns ± 1%     222ns ± 1%   -57.89%  (p=0.008 n=5+5)
    StdParseIP/v4-8             94.5ns ± 2%    89.0ns ± 1%    -5.78%  (p=0.008 n=5+5)
    StdParseIP/v6-8              150ns ± 1%     158ns ± 2%    +5.21%  (p=0.008 n=5+5)
    StdParseIP/v6_ellipsis-8     116ns ± 3%     125ns ± 2%    +7.59%  (p=0.008 n=5+5)
    StdParseIP/v6_v4-8           190ns ± 1%     183ns ± 2%    -3.89%  (p=0.008 n=5+5)
    StdParseIP/v6_zone-8         138ns ± 2%     133ns ± 2%    -3.20%  (p=0.024 n=5+5)
    
    name                      old alloc/op   new alloc/op   delta
    ParseIP/v4-8                 0.00B          0.00B           ~     (all equal)
    ParseIP/v6-8                 16.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
    ParseIP/v6_ellipsis-8        16.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
    ParseIP/v6_v4-8              32.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
    ParseIP/v6_zone-8            64.0B ± 0%     16.0B ± 0%   -75.00%  (p=0.008 n=5+5)
    StdParseIP/v4-8              16.0B ± 0%     16.0B ± 0%      ~     (all equal)
    StdParseIP/v6-8              16.0B ± 0%     16.0B ± 0%      ~     (all equal)
    StdParseIP/v6_ellipsis-8     16.0B ± 0%     16.0B ± 0%      ~     (all equal)
    StdParseIP/v6_v4-8           32.0B ± 0%     32.0B ± 0%      ~     (all equal)
    StdParseIP/v6_zone-8         16.0B ± 0%     16.0B ± 0%      ~     (all equal)
    
    name                      old allocs/op  new allocs/op  delta
    ParseIP/v4-8                  0.00           0.00           ~     (all equal)
    ParseIP/v6-8                  1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
    ParseIP/v6_ellipsis-8         1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
    ParseIP/v6_v4-8               2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
    ParseIP/v6_zone-8             4.00 ± 0%      1.00 ± 0%   -75.00%  (p=0.008 n=5+5)
    StdParseIP/v4-8               1.00 ± 0%      1.00 ± 0%      ~     (all equal)
    StdParseIP/v6-8               1.00 ± 0%      1.00 ± 0%      ~     (all equal)
    StdParseIP/v6_ellipsis-8      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
    StdParseIP/v6_v4-8            2.00 ± 0%      2.00 ± 0%      ~     (all equal)
    StdParseIP/v6_zone-8          1.00 ± 0%      1.00 ± 0%      ~     (all equal)
```