@danderson 2020-12-27
Still a draft because I need to fix up the error handling and comments, but opening the PR early for a question: after a bunch of false starts implementing a fancy parseIPv6, I took the implementation from the Go net package and adapted it. It's changed a fair bit, but the basic structure remains the same. How should we handle the licensing for that? Put it in a separate file and acknowledge the Go license for that?