@danderson 2020-12-28
Ready for review. Parsers return nice errors now, and are still fast (except the v6+zone case, I still don't know how it manages to be 2x slower than stdlib)